+++ src/helpers/cleanChildNodes.ts
@@ -1,5 +1,5 @@
 type CleanChildNodes = (node: HTMLElement) => void
 
-const cleanChildNodes: CleanChildNodes = node => node.childNodes.forEach((el: any) => el.remove())
+const cleanChildNodes: CleanChildNodes = node => node.childNodes.forEach(el => el.remove())
 
 export { cleanChildNodes }