+++ src/modes/default.ts
@@ -2,7 +2,7 @@ import { typingInterval } from '@svelte-typewriter/helpers'
 import type { TypewriterEffectFn, TypewriterModeFn } from '@svelte-typewriter/types'
 
 // Unify the function below with the one on default mode
-const typewriterEffect: TypewriterEffectFn = async ({ currentNode, text }: any, options: any) => {
+const typewriterEffect: TypewriterEffectFn = async ({ currentNode, text }, options) => {
 	currentNode.classList.add('typing')
 	for (let index = 0; index <= text.length; index++) {
 		const char = text[index]
@@ -15,23 +15,22 @@ const typewriterEffect: TypewriterEffectFn = async ({ currentNode, text }: any, 
 		: currentNode.classList.remove('typing')
 }
 
-const defaultMode: TypewriterModeFn = async ({ elements }: any, options: any) => {
+const defaultMode: TypewriterModeFn = async ({ elements }, options) => {
 	for (const element of elements) typewriterEffect(element, options)
 
 	// Modularize this
-	const { currentNode: lastElementToFinish } = elements.reduce(
-		(longestTextElement: any, element: any) => {
-			const longestTextLength = longestTextElement.text.length
-			return element.text.length > longestTextLength
-				? (longestTextElement = element)
-				: longestTextElement
-		}
-	)
+	const { currentNode: lastElementToFinish } = elements.reduce((longestTextElement, element) => {
+		const longestTextLength = longestTextElement.text.length
+		return element.text.length > longestTextLength
+			? (longestTextElement = element)
+			: longestTextElement
+	})
 
 	const observer = new MutationObserver(mutations => {
 		mutations.forEach((mutation: any) => {
+			const elementAttributeChanged = mutation.type === 'attributes'
 			const lastElementFinishedTyping = !mutation.target.classList.contains('typing')
-			if (mutation.type === 'attributes' && lastElementFinishedTyping) {
+			if (elementAttributeChanged && lastElementFinishedTyping) {
 				options.dispatch('done')
 			}
 		})