+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './scramble'
 export * from './typewriter'
 export * from './loopRandom'
-export * from './loop'