+++ src/types/loop.ts
@@ -1,3 +1,0 @@
-type CleanChildNodes = (node: HTMLElement) => void
-
-export { CleanChildNodes }